Factoring out sequence worker state
