+++ crates/sui-simple-fullnode/Cargo.toml
@@ -12,6 +12,7 @@ anyhow = { version = "1.0.64", features = ["backtrace"] }
 git-version = "0.3.5"
 const-str = "0.5.3"
 futures = "0.3.23"
+prometheus = "0.13"
 tracing = "0.1.36"
